The casino code for major wins (in all the different slot types) was missing a cast to integer, this resulted in a traceback on Python 3.x.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/maps@11481 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ trunk/python/casino/diamondslots.py
@@ -65,7 +65,7 @@ if (object) and not object.Unpaid:
                 pay = slotmajor[slotlist.index(item)]
              else:
                 break
-             payoff = pot*pay
+             payoff = int(pot*pay)
              Slots.payoff(payoff)
              id = Crossfire.CreateObjectByName(cointype)
              id.Quantity = payoff
